+++ examples/blinky.rs
@@ -1,62 +1,39 @@
 //! Blinks an LED
 #![deny(unsafe_code)]
 #![deny(warnings)]
-#![feature(proc_macro)]
 #![no_std]
 
-extern crate f3;
 extern crate cortex_m;
-extern crate cortex_m_rtfm as rtfm;
+extern crate f3;
 
-use cortex_m::peripheral::SystClkSource;
-use f3::led::{self, LEDS};
-use rtfm::{app, Threshold};
+use f3::hal::delay::Delay;
+use f3::hal::prelude::*;
+use f3::hal::stm32f30x;
+use f3::led::Led;
 
-// CONFIGURATION
-const FREQUENCY: u32 = 4; // Hz
+fn main() {
+    let cp = cortex_m::Peripherals::take().unwrap();
+    let dp = stm32f30x::Peripherals::take().unwrap();
 
-// TASKS & RESOURCES
-app! {
-    device: f3::stm32f30x,
+    let mut flash = dp.FLASH.constrain();
+    let mut rcc = dp.RCC.constrain();
+    let mut gpioe = dp.GPIOE.split(&mut rcc.ahb);
 
-    resources: {
-        static ON: bool = false;
-    },
+    // clock configuration using the default settings (all clocks run at 8 MHz)
+    let clocks = rcc.cfgr.freeze(&mut flash.acr);
+    // TRY this alternate clock configuration (all clocks run at 16 MHz)
+    // let clocks = rcc.cfgr.sysclk(16.mhz()).freeze(&mut flash.acr);
 
-    tasks: {
-        SYS_TICK: {
-            path: toggle,
-            resources: [ON],
-        },
-    },
-}
+    let mut led: Led = gpioe
+        .pe9
+        .into_push_pull_output(&mut gpioe.moder, &mut gpioe.otyper)
+        .into();
+    let mut delay = Delay::new(cp.SYST, clocks);
 
-// INITIALIZATION PHASE
-fn init(p: init::Peripherals, _r: init::Resources) {
-    led::init(p.GPIOE, p.RCC);
-
-    p.SYST.set_clock_source(SystClkSource::Core);
-    p.SYST.set_reload(8_000_000 / FREQUENCY);
-    p.SYST.enable_interrupt();
-    p.SYST.enable_counter();
-}
-
-// IDLE LOOP
-fn idle() -> ! {
-    // Sleep
     loop {
-        rtfm::wfi();
-    }
-}
-
-// TASKS
-// Toggle the state of the LED
-fn toggle(_t: &mut Threshold, r: SYS_TICK::Resources) {
-    **r.ON = !**r.ON;
-
-    if **r.ON {
-        LEDS[0].on();
-    } else {
-        LEDS[0].off();
+        led.on();
+        delay.delay_ms(1_000_u16);
+        led.off();
+        delay.delay_ms(1_000_u16);
     }
 }