+++ src/lib.rs
@@ -1,49 +1,50 @@
-//! Board Support Crate for the STM32F3DISCOVERY
+//! Board support crate for the STM32F3DISCOVERY
 //!
 //! # Usage
 //!
-//! Follow `cortex-m-quickstart` [instructions][i] but remove the `memory.x`
-//! linker script and the `build.rs` build script file as part of the
-//! configuration of the quickstart crate. Additionally, uncomment the "if using
-//! ITM" block in the `.gdbinit` file.
+//! ``` text
+//! $ # if you don't have the clone subcommand
+//! $ cargo install cargo-clone
 //!
-//! [i]: https://docs.rs/cortex-m-quickstart/0.2.0/cortex_m_quickstart/
+//! $ cargo clone f3 --vers 0.5.0
+//!
+//! # on another terminal
+//! $ openocd -f interface/stlink-v2-1.cfg -f target/stm32f3x.cfg
+//!
+//! # flash and debug the "Hello, world" example
+//! $ cd f3
+//! $ xargo run --example hello
+//! ```
+//!
+//! You'll need to have both OpenOCD and arm-none-eabi-ld installed.
 //!
 //! # Examples
 //!
-//! Check the [examples] module.
+//! See the [examples] module.
 //!
-//! [examples]: ./examples/index.html
+//! [examples]: examples/index.html
 
 #![deny(missing_docs)]
 #![deny(warnings)]
-#![feature(const_fn)]
-#![feature(const_unsafe_cell_new)]
-#![feature(const_cell_new)]
-#![feature(get_type_id)]
-#![feature(never_type)]
-#![feature(unsize)]
 #![no_std]
 
-extern crate cast;
-extern crate embedded_hal as hal;
-extern crate nb;
-extern crate static_ref;
+pub extern crate l3gd20;
+pub extern crate lsm303dlhc;
+pub extern crate stm32f30x_hal as hal;
 
-pub extern crate stm32f30x;
+use hal::i2c::I2c;
+use hal::spi::Spi;
+use hal::stm32f30x::{I2C1, SPI1};
+use hal::gpio::{AF4, AF5, Output, PushPull};
+use hal::gpio::gpioa::{PA5, PA6, PA7};
+use hal::gpio::gpiob::{PB6, PB7};
+use hal::gpio::gpioe::PE3;
 
-// For documentation only
 pub mod examples;
-
-pub mod dma;
 pub mod led;
-pub mod serial;
-pub mod timer;
-pub mod time;
 
-pub mod frequency;
-use frequency::*;
+/// On board L3GD20 connected to the SPI1 bus via the pins PA5, PA6, PA7 and PE3
+pub type L3gd20 = l3gd20::L3gd20<Spi<SPI1, (PA5<AF5>, PA6<AF5>, PA7<AF5>)>, PE3<Output<PushPull>>>;
 
-pub use hal::prelude;
-pub use serial::Serial;
-
+/// On board LSM303DLHC connected to the I2C1 bus via the PB6 and PB7 pins
+pub type Lsm303dlhc = lsm303dlhc::Lsm303dlhc<I2c<I2C1, (PB6<AF4>, PB7<AF4>)>>;